some progress
Here I'm trying to expand into encrypt and decrypt to work both with the
remote service and the local NodeJS API. Not only that, but I'm also
trying to make it work as part of our build bundle.

I was successful to call encrypt and decrypt with the live services,
but once I added the code related to Node crypto, I started finding
problems. I was not successful to build it with the local node
implementation (so I also wasn't able to finish testing it). The last
error I'm getting is that apparently "crypto.constants" does not exist
in @types/node

I had this error in my local tests. I was able to fix it by upgrading
@types/node to ^12.6.1. We're currenlty using 8.0.0. Would it be ok to
upgrade this package?